RSDK-13481: Remove unsafe usage of `robotutils.Connect`. (#5815)
